+++ flap-annotations/src/main/java/me/yifeiyuan/flap/annotations/Flap.java
@@ -1,6 +1,5 @@
 package me.yifeiyuan.flap.annotations;
 
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
@@ -10,6 +9,13 @@ import java.lang.annotation.Target;
 @Target(ElementType.TYPE)
 public @interface Flap {
 
+    /**
+     * @return the layout res id
+     */
     int layoutId() default -1;
 
+    /**
+     * @return true then FlapProcessor will auto register the Factory.
+     */
+    boolean autoRegister() default true;
 }